requirements.txt
flask
gunicorn
flask_cors
google-cloud-vision
google-cloud-storage
werkzeug
